"apply replaygain" option uses locales when it should not
The replaygain tags inside a file are in the format:
REPLAYGAIN_TRACK_GAIN=-2.29 dB
REPLAYGAIN_TRACK_PEAK=0.96014404

$ LC_NUMERIC=sv_SE.utf8 flac -d --apply-replaygain-which-is-not-lossless *.flac

01 d-moll.flac: INFO: applying album ReplayGain (gain=-2,00dB+preamp=0,0dB, low noise shaping, hard limiting) to output
01 d-moll.flac: WARNING: applying ReplayGain is not lossless
01 d-moll.flac: done

$ LC_NUMERIC=C flac -d --apply-replaygain-which-is-not-lossless *.flac

01 d-moll.flac: INFO: applying album ReplayGain (gain=-2.29dB+preamp=0.0dB, low noise shaping, hard limiting) to output
01 d-moll.flac: WARNING: applying ReplayGain is not lossless
01 d-moll.flac: done

Note that the gains differ. sv_SE uses a decimal comma while the data in
the tags use a decimal point. In that locale, the decoder considers the
decimal point an incorrect character and truncates the tag to "-2 dB".

Sourceforge issue #3288157.